+++ webots_ros2_driver/include/webots_ros2_driver/plugins/static/Ros2Receiver.hpp
@@ -4,11 +4,9 @@
 #include <webots_ros2_driver/WebotsNode.hpp>
 #include <webots_ros2_driver/plugins/Ros2SensorPlugin.hpp>
 
+#include <geometry_msgs/msg/vector3_stamped.hpp>
+#include <webots_ros2_msgs/msg/float_stamped.hpp>
 #include <webots_ros2_msgs/msg/string_stamped.hpp>
-#include <webots_ros2_msgs/srv/get_float.hpp>
-#include <webots_ros2_msgs/srv/get_int.hpp>
-#include <webots_ros2_msgs/srv/receiver_get_emitter_direction.hpp>
-#include <webots_ros2_msgs/srv/set_int.hpp>
 
 namespace webots_ros2_driver {
   class Ros2Receiver : public Ros2SensorPlugin {
@@ -20,23 +18,14 @@ namespace webots_ros2_driver {
     bool mustPublish();
     void publishData();
 
-    void enable_callback(const std::shared_ptr<webots_ros2_msgs::srv::SetInt::Request> request,
-                         std::shared_ptr<webots_ros2_msgs::srv::SetInt::Response> response);
-    void get_emitter_direction_callback(
-      const std::shared_ptr<webots_ros2_msgs::srv::ReceiverGetEmitterDirection::Request> request,
-      std::shared_ptr<webots_ros2_msgs::srv::ReceiverGetEmitterDirection::Response> response);
-    void get_sampling_period_callback(const std::shared_ptr<webots_ros2_msgs::srv::GetInt::Request> request,
-                                      std::shared_ptr<webots_ros2_msgs::srv::GetInt::Response> response);
-    void get_signal_strength_callback(const std::shared_ptr<webots_ros2_msgs::srv::GetFloat::Request> request,
-                                      std::shared_ptr<webots_ros2_msgs::srv::GetFloat::Response> response);
     // ROS2 topics
     rclcpp::Publisher<webots_ros2_msgs::msg::StringStamped>::SharedPtr mDataPublisher;
     webots_ros2_msgs::msg::StringStamped mDataMessage;
+    rclcpp::Publisher<webots_ros2_msgs::msg::FloatStamped>::SharedPtr mSignalPublisher;
+    webots_ros2_msgs::msg::FloatStamped mSignalMessage;
+    rclcpp::Publisher<geometry_msgs::msg::Vector3Stamped>::SharedPtr mDirectionPublisher;
+    geometry_msgs::msg::Vector3Stamped mDirectionMessage;
 
-    rclcpp::Service<webots_ros2_msgs::srv::SetInt>::SharedPtr enable_service_;
-    rclcpp::Service<webots_ros2_msgs::srv::ReceiverGetEmitterDirection>::SharedPtr get_emitter_direction_service_;
-    rclcpp::Service<webots_ros2_msgs::srv::GetInt>::SharedPtr get_sampling_period_service_;
-    rclcpp::Service<webots_ros2_msgs::srv::GetFloat>::SharedPtr get_signal_strength_service_;
     // Device
     WbDeviceTag mReceiver;
     // Runtime vars