+++ webots_ros2_tests/CHANGELOG.rst
@@ -2,10 +2,15 @@
 Changelog for package webots_ros2_tests
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-2023.0.3 (2023-XX-XX)
+2023.0.4 (2023-XX-XX)
+------------------
+* Added support for painted point clouds
+
+2023.0.3 (2023-04-12)
 ------------------
 * Improved the TIAGo test by publishing the initial position when navigation is ready.
 * Fixed and improved Turtlebot navigation and mapping tests.
+* Fixed and improved e-puck navigation test.
 
 2022.1.4 (2022-11-18)
 ------------------